simplify folder structure
