+++ src/api/app.py
@@ -14,13 +14,10 @@ from fastapi.middleware.cors import CORSMiddleware
 from dotenv import load_dotenv
 import uvicorn
 
-from agents.conversation_agent_factory import ConversationAgentFactory
-from agents.search_agent_factory import SearchAgentFactory
-from agents.sql_agent_factory import SQLAgentFactory
-from agents.chart_agent_factory import ChartAgentFactory
-from api.api_routes import router as backend_router
-from api.history_routes import router as history_router
-from api.history_sql_routes import router as history_sql_router
+from agent_factory import AgentFactory, AgentType
+from chat import router as chat_router
+from history import router as history_router
+from history_sql import router as history_sql_router
 
 load_dotenv()
 
@@ -33,15 +30,12 @@ async def lifespan(fastapi_app: FastAPI):
     On startup, initializes the Azure AI agent using the configuration and attaches it to the app state.
     On shutdown, deletes the agent instance and performs any necessary cleanup.
     """
-    fastapi_app.state.agent = await ConversationAgentFactory.get_agent()
-    fastapi_app.state.search_agent = await SearchAgentFactory.get_agent()
-    fastapi_app.state.sql_agent = await SQLAgentFactory.get_agent()
-    fastapi_app.state.chart_agent = await ChartAgentFactory.get_agent()
+    fastapi_app.state.agent = await AgentFactory.get_agent(AgentType.CONVERSATION)
+    fastapi_app.state.search_agent = await AgentFactory.get_agent(AgentType.SEARCH)
+    fastapi_app.state.sql_agent = await AgentFactory.get_agent(AgentType.SQL)
+    fastapi_app.state.chart_agent = await AgentFactory.get_agent(AgentType.CHART)
     yield
-    await ConversationAgentFactory.delete_agent()
-    await SearchAgentFactory.delete_agent()
-    await SQLAgentFactory.delete_agent()
-    await ChartAgentFactory.delete_agent()
+    await AgentFactory.delete_all_agents()
     fastapi_app.state.sql_agent = None
     fastapi_app.state.search_agent = None
     fastapi_app.state.agent = None
@@ -67,7 +61,7 @@ def build_app() -> FastAPI:
     )
 
     # Include routers
-    fastapi_app.include_router(backend_router, prefix="/api", tags=["backend"])
+    fastapi_app.include_router(chat_router, prefix="/api", tags=["chat"])
     fastapi_app.include_router(history_router, prefix="/history", tags=["history"])
     fastapi_app.include_router(history_sql_router, prefix="/historyfab", tags=["historyfab"])
