--- src/api/common/database/sqldb_service.py
@@ -1,114 +1,0 @@
-# from datetime import datetime
-import datetime
-import struct
-
-# import pandas as pd
-# from api.models.input_models import ChartFilters
-from common.config.config import Config
-import logging
-from azure.identity.aio import DefaultAzureCredential
-import pyodbc
-
-
-async def get_db_connection():
-    """Get a connection to the SQL database"""
-    config = Config()
-
-    server = config.sqldb_server
-    database = config.sqldb_database
-    username = config.sqldb_username
-    password = config.sqldb_database
-    driver = config.driver
-    mid_id = config.mid_id
-
-    try:
-        async with DefaultAzureCredential(managed_identity_client_id=mid_id) as credential:
-            token = await credential.get_token("https://database.windows.net/.default")
-            token_bytes = token.token.encode("utf-16-LE")
-            token_struct = struct.pack(
-                f"<I{len(token_bytes)}s",
-                len(token_bytes),
-                token_bytes
-            )
-            SQL_COPT_SS_ACCESS_TOKEN = 1256
-
-            # Set up the connection
-            connection_string = f"DRIVER={driver};SERVER={server};DATABASE={database};"
-            conn = pyodbc.connect(
-                connection_string, attrs_before={SQL_COPT_SS_ACCESS_TOKEN: token_struct}
-            )
-
-            logging.info("Connected using Default Azure Credential")
-            return conn
-    except pyodbc.Error as e:
-        logging.error("Failed with Default Credential: %s", str(e))
-        conn = pyodbc.connect(
-            f"DRIVER={driver};SERVER={server};DATABASE={database};UID={username};PWD={password}",
-            timeout=5)
-
-        logging.info("Connected using Username & Password")
-        return conn
-
-
-async def execute_sql_query(sql_query):
-    """
-    Executes a given SQL query and returns the result as a concatenated string.
-    """
-    conn = await get_db_connection()
-    cursor = None
-    try:
-        cursor = conn.cursor()
-        cursor.execute(sql_query)
-        result = ''.join(str(row) for row in cursor.fetchall())
-        return result
-    except Exception as e:
-        logging.error("Error executing SQL query: %s", e)
-        return None
-    finally:
-        if cursor:
-            cursor.close()
-        conn.close()
-        
-
-async def adjust_processed_data_dates():
-    """
-    Adjusts the dates in the processed_data, km_processed_data, and processed_data_key_phrases tables
-    to align with the current date.
-    """
-    conn = await get_db_connection()
-    cursor = None
-    try:
-        cursor = conn.cursor()
-        # Adjust the dates to the current date
-        today = datetime.today()
-        cursor.execute(
-            "SELECT MAX(CAST(StartTime AS DATETIME)) FROM [dbo].[processed_data]"
-        )
-        max_start_time = (cursor.fetchone())[0]
-
-        if max_start_time:
-            days_difference = (today - max_start_time).days - 1
-            if days_difference != 0:
-                # Update processed_data table
-                cursor.execute(
-                    "UPDATE [dbo].[processed_data] SET StartTime = FORMAT(DATEADD(DAY, ?, StartTime), 'yyyy-MM-dd "
-                    "HH:mm:ss'), EndTime = FORMAT(DATEADD(DAY, ?, EndTime), 'yyyy-MM-dd HH:mm:ss')",
-                    (days_difference, days_difference)
-                )
-                # Update km_processed_data table
-                cursor.execute(
-                    "UPDATE [dbo].[km_processed_data] SET StartTime = FORMAT(DATEADD(DAY, ?, StartTime), 'yyyy-MM-dd "
-                    "HH:mm:ss'), EndTime = FORMAT(DATEADD(DAY, ?, EndTime), 'yyyy-MM-dd HH:mm:ss')",
-                    (days_difference, days_difference)
-                )
-                # Update processed_data_key_phrases table
-                cursor.execute(
-                    "UPDATE [dbo].[processed_data_key_phrases] SET StartTime = FORMAT(DATEADD(DAY, ?, StartTime), "
-                    "'yyyy-MM-dd HH:mm:ss')", (days_difference,)
-                )
-                # Commit the changes
-                conn.commit()
-    finally:
-        if cursor:
-            cursor.close()
-        conn.close()
