--- src/api/helpers/streaming_helper.py
@@ -1,11 +1,0 @@
-import logging
-
-
-async def stream_processor(response):
-    try:
-        async for message in response:
-            if message.content:
-                yield message.content
-    except Exception as e:
-        logging.error(f"Error processing streaming response: {e}", exc_info=True)
-        raise
